pyproject.toml
[tool.poetry]
name = "assurity-demo"
version = "0.1.0"
description = "An automated pipeline for OCR processing of insurance claims documents"
authors = ["bartosz.roguski@bluelabellabs.com"]
readme = "README.md"
packages = [{include = "assurity_demo", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
opencv-python = "^4.7.0"
pillow = "^10.0.0"
openai = "^1.1.0"
python-dotenv = "^1.0.0"
streamlit = "^1.44.1"
langchain = "^0.3.23"
langchain-openai = "^0.3.12"
dynaconf = "^3.2.10"
pydantic-settings = "^2.8.1"
levenshtein = "^0.27.1"
sentence-transformers = ">=3.0.0"
scikit-learn = "^1.6.1"
fire = "^0.7.0"
logzero = "^1.7.0"
scikit-image = "^0.25.2"
imagehash = "^4.3.2"
pypdf2 = "^3.0.1"
pdf2image = "^1.17.0"
tqdm = "^4.67.1"
langchain-google-genai = "^2.1.3"
llama-parse = "0.6.4.post1"
promptlayer = "^1.0.48"
langchain-community = "^0.3.23"
tabulate = "^0.9.0"
pypdf = "^5.5.0"
transformers = ">=4.45.0"

[tool.poetry.group.dev.dependencies]
ipython = "^9.0.2"
pre-commit = "^4.2.0"
ruff = "^0.11.4"
jupyter = "^1.1.1"
ipykernel = "^6.29.5"
seaborn = "^0.13.2"
isort = "^6.0.1"
ipdb = "^0.13.13"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "B", # flake8-bugbear
]

[tool.ruff.lint.isort]
length-sort = true
lines-between-types = 1


[tool.mypy]
python_version = "3.13"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 